core/gemini-connector/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.8.2
pydantic-settings==2.1.0
requests==2.31.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
google-generativeai==0.3.2
python-dotenv==1.0.0
